finished util and reducers
