+++ frontend/src/actions/session_actions.js
@@ -48,6 +48,6 @@ export const login = user => dispatch =>
 
 export const logout = () => dispatch => {
     localStorage.removeItem('jwtToken')
-    APIUtil.setAuthToken(false)
+    SessionAPIUtil.setAuthToken(false)
     dispatch(logoutUser())
 };
