+++ frontend/src/reducers/root_reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import session from './session_reducer';
+import session from './session_api_reducer';
 import errors from './errors_reducer';
 
 const RootReducer = combineReducers({
